more

+++ testset/dispbmp/windows/comptbmp/lowercase.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 for file in *.{ico,ICO,rc,RC} RESOURCE.H resource.h; do
-	lc=`echo "$file" | tr '[:upper:]' '[:lower:]'`
-	if [ "$file" == "$lc" ]; then
-		true
-	else
-		mv -v "$file" "$lc" || exit 1
+	if [ -f "$file" ]; then
+		lc=`echo "$file" | tr '[:upper:]' '[:lower:]'`
+		if [ "$file" == "$lc" ]; then
+			true
+		else
+			mv -v "$file" "$lc" || exit 1
+		fi
 	fi
 done
